[REFATORAÇÃO] - Separação de métodos nos testes da carteira 138 Itau e organização do teste na pasta do banco
